If file array was provided, iterate through files to delete them
This is done to ensure that the file array is not undefined before
iteration.

+++ middleware/helpers.js
@@ -27,7 +27,9 @@ exports.deleteFile = (filePath) => {
 }
 
 exports.deleteFiles = (fileArray) => {
-    for (let file of fileArray) {
-        exports.deleteFile(file.path);
+    if (fileArray) {
+        for (let file of fileArray) {
+            exports.deleteFile(file.path);
+        }
     }
 }
